requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
filelock==3.0.12
huggingface-hub==0.0.16
idna==3.2
joblib==1.0.1
mecab-python3==1.0.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pyparsing==2.4.7
pytextspan==0.5.4
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
sacremoses==0.0.45
sentencepiece==0.1.96
six==1.16.0
tokenizers==0.10.3
torch==1.9.0
tqdm==4.62.2
transformers==4.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
unidic
unidic-lite
torchtext
ipadic
matplotlib
umap-learn
flake8
black
isort
